+++ backend/app/models/symbol.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Float, Integer, String, func
+from sqlalchemy import BigInteger, Boolean, DateTime, Float, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.database import Base
@@ -17,7 +17,7 @@ class Symbol(Base):
     name: Mapped[str] = mapped_column(String(255), nullable=False, default="")
     exchange: Mapped[str] = mapped_column(String(20), nullable=False, default="NASDAQ")
     last_price: Mapped[float] = mapped_column(Float, nullable=True)
-    avg_volume: Mapped[int] = mapped_column(Integer, nullable=True)
+    avg_volume: Mapped[int] = mapped_column(BigInteger, nullable=True)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
     updated_at: Mapped[datetime] = mapped_column(
